Add files via upload
Cleaned up the lcd screen issues using sprintf and a print buffer for the two line display.  Is working with no problems to the lcd screen becoming corrupted.  When you download the sketch uncomment the screen rest in the loop section and the backlight will go off after a set time and come back on with a button press of the select button.  enjoy